move diffing tests to test/utils/diffing, and force a full recheck on tag removal
