+++ test/utils/diffing/getFenceDiffTests.ts
@@ -1,5 +1,5 @@
-import type NormalizedPath from '../../src/types/NormalizedPath';
-import { diffList, getFenceDiff } from '../../src/utils/diffing/getFenceDiff';
+import type NormalizedPath from '../../../src/types/NormalizedPath';
+import { diffList, getFenceDiff } from '../../../src/utils/diffing/getFenceDiff';
 
 let strictCmp = (a: any, b: any): boolean => {
     return a === b;
